+++ frontend/src/store/reducers.tsx
@@ -2,6 +2,7 @@ import { connectRouter } from 'connected-react-router';
 import { BrowserHistory } from 'history';
 import { combineReducers } from 'redux';
 import { reducersAuth } from './auth/reducers';
+import { reducersBudgets } from './budgets/reducers';
 import { reducersMisc } from './misc/reducers';
 
 export const createRootReducer = (history: BrowserHistory) => {
@@ -11,5 +12,6 @@ export const createRootReducer = (history: BrowserHistory) => {
     register: reducersAuth.register,
     restCheck: reducersMisc.restCheck,
     expense: reducersMisc.expenseIncrement,
+    budgets: reducersBudgets.budgets,
   });
 };