+++ docs/CHANGELOG.md
@@ -7,20 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [[0.4.3]] - 2022-03-24
+## [[0.4.4]] - 2022-03-25
 
-###### _([diff: 0.4.1-0.4.2])_
+###### _([diff: 0.4.3-0.4.4])_
+
+### Added
+
+- Getting All User Budgets (dispatched via _redux_ store)
+    - New endpoint: `/rest/api/budget/all` _(get)_
+
+## [[0.4.3]] - 2022-03-25
+
+###### _([diff: 0.4.2-0.4.3])_
 
 ### Added
 
 - Filtering & Display for `Budget` in Django Admin View
 - Adding new Budget via API Call (integrated)
+    - New endpoint: `/rest/api/budget/` _(post)_
 
 ### To-do:
 
 - Handle new budget API call response on frontend.
 
-## [[0.4.2]] - 2022-03-24
+## [[0.4.2]] - 2022-03-25
 
 ###### _([diff: 0.4.1-0.4.2])_
 
@@ -37,7 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Still no tests & CI pipelines 🤡_
 
-## [[0.4.1]] - 2022-03-24
+## [[0.4.1]] - 2022-03-25
 
 ###### _([diff: 0.4.0-0.4.1])_
 
@@ -45,7 +55,7 @@ _Still no tests & CI pipelines 🤡_
 
 - Budget Modal Form
 
-## [[0.4.0]] - 2022-03-24
+## [[0.4.0]] - 2022-03-25
 
 ###### _([diff: 0.4.0-alpha-0.4.0])_
 
@@ -240,6 +250,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.4.3-0.4.4]: https://github.com/Luzkan/FamilyBudget/compare/0.4.3...0.4.4
 [diff: 0.4.2-0.4.3]: https://github.com/Luzkan/FamilyBudget/compare/0.4.2...0.4.3
 [diff: 0.4.1-0.4.2]: https://github.com/Luzkan/FamilyBudget/compare/0.4.1...0.4.2
 [diff: 0.4.0-0.4.1]: https://github.com/Luzkan/FamilyBudget/compare/0.4.0...0.4.1
@@ -256,6 +267,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.4.4]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.4
 [0.4.3]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.3
 [0.4.2]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.2
 [0.4.1]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.1